+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - start : fix hostname warning
 - start (#046) : run the jail in a clean enviroment
 - term (#046) : spawn the shell in a jail using the jailed user environment
+- add-fscomp (#045) : check the mount point and create it, if missing
 
 ## [0.5.7] 2018-06-28
 ### Added